kernel: Revert "PCI: imx6: Skip controller_id generation logic for i.MX7D"
Issue: LIN1024-6852

The stable commit 2a12efc567a ("PCI: imx6: Skip controller_id generation
logic for i.MX7D") introduced a PCIe regression on the i.MX7 board. Our
preliminary investigation indicates that the issue stems from missing
dependent commits in 2a12efc567a. We plan to work with the upstream stable
team to resolve this.

However, as this regression causes the i.MX7 boot process to hang, and
given that a fix may take considerable time to be merged into the stable
kernel, we have decided to temporarily revert this commit and maintain
the revert as a local patch in the BSP layer. This workaround will be
removed once the issue is resolved in upstream stable kernel.

Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>